edits

+++ src/cub_init.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   cub_init.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mlagrini <mlagrini@student.1337.ma>        +#+  +:+       +#+        */
+/*   By: hrahmane <hrahmane@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/04 22:03:21 by mlagrini          #+#    #+#             */
-/*   Updated: 2023/09/24 10:45:15 by mlagrini         ###   ########.fr       */
+/*   Updated: 2023/09/24 13:27:22 by hrahmane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,8 +49,8 @@ int	run_mlx(t_cub *var)
 {
 	int			color;
 
-	var->x = 0;
-	var->y = 0;
+	var->x = 0.0;
+	var->y = 0.0;
 	color = 0;
 	var->mlx = mlx_init(WIDTH, HEIGHT, "test", false);
 	var->img = mlx_new_image(var->mlx, WIDTH, HEIGHT);
@@ -67,14 +67,14 @@ int	run_mlx(t_cub *var)
 
 void	init_variables(t_cub *var)
 {
-	var->p.radius = 3;
+	var->p.radius = 3.0;
 	var->x_max *= T_SIZE;
 	var->y_max *= T_SIZE;
-	var->p.radians = M_PI / 180;
+	var->p.radians = (float)( M_PI / 180);
 	var->p.fov = 60 * var->p.radians;
 	var->p.rot_speed = 0.07;
-	var->p.move_speed = 5;
-	var->p.rotation_angle = M_PI / 2;
+	var->p.move_speed = 5.0;
+	var->p.rotation_angle = (float)(M_PI / 2);
 	get_direction(var);
 }
 
@@ -103,13 +103,6 @@ int	init_vars(t_cub *var, int ac, char **av)
 		return (free_phase1(var), 1);
 	collect_paths(var);
 	get_tex(var);
-	// get_images(path);
-	// var->txt = mlx_load_png(var->north[1]);
-	// if (!var->txt)
-	// {
-	// 	printf("can't open file\n");
-	// 	exit(1);
-	// }
 	init_variables(var);
 	return (0);
 }
